+++ src/Api/Api.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Copyright 2020 Cloud Creativity Limited
  *
@@ -25,17 +24,14 @@ use CloudCreativity\LaravelJsonApi\Codec\EncodingList;
 use CloudCreativity\LaravelJsonApi\Contracts\Client\ClientInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\ContainerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Encoder\SerializerInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Repositories\ErrorRepositoryInterface;
+use CloudCreativity\LaravelJsonApi\Contracts\Exceptions\ExceptionParserInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Resolver\ResolverInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Store\StoreInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Validators\ValidatorFactoryInterface;
 use CloudCreativity\LaravelJsonApi\Factories\Factory;
 use CloudCreativity\LaravelJsonApi\Http\Responses\Responses;
 use CloudCreativity\LaravelJsonApi\Resolver\AggregateResolver;
 use CloudCreativity\LaravelJsonApi\Resolver\NamespaceResolver;
 use GuzzleHttp\Client;
-use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
-use Neomerx\JsonApi\Contracts\Encoder\EncoderInterface;
 use Neomerx\JsonApi\Contracts\Http\Headers\MediaTypeInterface;
 use Neomerx\JsonApi\Contracts\Http\Headers\SupportedExtensionsInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
@@ -74,12 +70,6 @@ class Api
     private $decodings;
 
     /**
-     * @var array
-     * @deprecated 2.0.0
-     */
-    private $errors;
-
-    /**
      * @var bool
      */
     private $useEloquent;
@@ -110,17 +100,6 @@ class Api
     private $store;
 
     /**
-     * @var CodecMatcherInterface|null
-     */
-    private $codecMatcher;
-
-    /**
-     * @var ErrorRepositoryInterface|null
-     * @deprecated 2.0.0
-     */
-    private $errorRepository;
-
-    /**
      * @var Responses|null
      */
     private $responses;
@@ -141,6 +120,11 @@ class Api
     private $transactions;
 
     /**
+     * @var string|null
+     */
+    private $modelNamespace;
+
+    /**
      * Api constructor.
      *
      * @param Factory $factory
@@ -152,10 +136,10 @@ class Api
      * @param Jobs $jobs
      * @param bool $useEloquent
      * @param string|null $supportedExt
-     * @param array $errors
      * @param array $providers
      * @param string|null $connection
      * @param bool $transactions
+     * @param string|null $modelNamespace
      */
     public function __construct(
         Factory $factory,
@@ -167,10 +151,10 @@ class Api
         Jobs $jobs,
         $useEloquent = true,
         $supportedExt = null,
-        array $errors = [],
         array $providers = [],
         string $connection = null,
-        bool $transactions = true
+        bool $transactions = true,
+        string $modelNamespace = null
     ) {
         $this->factory = $factory;
         $this->resolver = $resolver;
@@ -181,10 +165,10 @@ class Api
         $this->jobs = $jobs;
         $this->useEloquent = $useEloquent;
         $this->supportedExt = $supportedExt;
-        $this->errors = $errors;
         $this->providers = $providers;
         $this->connection = $connection;
         $this->transactions = $transactions;
+        $this->modelNamespace = $modelNamespace;
     }
 
     /**
@@ -192,10 +176,8 @@ class Api
      */
     public function __clone()
     {
-        $this->schemas = null;
+        $this->container = null;
         $this->store = null;
-        $this->codecMatcher = null;
-        $this->errorRepository = null;
     }
 
     /**
@@ -285,19 +267,6 @@ class Api
     }
 
     /**
-     * @return ErrorRepositoryInterface
-     * @deprecated 2.0.0
-     */
-    public function getErrors()
-    {
-        if (!$this->errorRepository) {
-            $this->errorRepository = $this->factory->createErrorRepository($this->errors);
-        }
-
-        return $this->errorRepository;
-    }
-
-    /**
      * @return SupportedExtensionsInterface|null
      */
     public function getSupportedExtensions()
@@ -354,17 +323,6 @@ class Api
     }
 
     /**
-     * Get the matched encoder, or a default encoder.
-     *
-     * @return EncoderInterface
-     * @deprecated 2.0.0 use `encoder` to create an encoder.
-     */
-    public function getEncoder()
-    {
-        return $this->encoder();
-    }
-
-    /**
      * Get the default database connection for the API.
      *
      * @return string|null
@@ -382,6 +340,23 @@ class Api
     public function hasTransactions(): bool
     {
         return $this->transactions;
+    }
+
+    /**
+     * @return ExceptionParserInterface
+     * @todo add this to config.
+     */
+    public function exceptions(): ExceptionParserInterface
+    {
+        return app(ExceptionParserInterface::class);
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getModelNamespace(): ?string
+    {
+        return $this->modelNamespace;
     }
 
     /**
@@ -468,18 +443,6 @@ class Api
     }
 
     /**
-     * @return ValidatorFactoryInterface
-     * @deprecated 2.0.0
-     */
-    public function validators()
-    {
-        return $this->factory->createValidatorFactory(
-            $this->getErrors(),
-            $this->getStore()
-        );
-    }
-
-    /**
      * Register a resource provider with this API.
      *
      * @param AbstractProvider $provider
@@ -488,7 +451,6 @@ class Api
     public function register(AbstractProvider $provider)
     {
         $this->resolver->attach($provider->getResolver());
-        $this->errors = array_replace($provider->getErrors(), $this->errors);
     }
 
 }