+++ src/Document/ResourceObject.php
@@ -17,8 +17,6 @@
 
 namespace CloudCreativity\LaravelJsonApi\Document;
 
-use CloudCreativity\LaravelJsonApi\Contracts\Object\ResourceObjectInterface;
-use CloudCreativity\LaravelJsonApi\Object\ResourceObject as LegacyResourceObject;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
@@ -669,33 +667,14 @@ class ResourceObject implements Arrayable, \IteratorAggregate, \JsonSerializable
     }
 
     /**
-     * Convert the object to a legacy resource object.
-     *
-     * @return ResourceObjectInterface
-     * @deprecated 2.0.0
-     */
-    public function toObject(): ResourceObjectInterface
-    {
-        $json = json_encode($this);
-
-        return new LegacyResourceObject(json_decode($json, false));
-    }
-
-    /**
      * @return Collection
      */
     private function fieldValues(): Collection
     {
-        $fields = collect($this->attributes)->merge($this->getRelations())->merge([
+        return collect($this->attributes)->merge($this->getRelations())->merge([
             'type' => $this->type,
             'id' => $this->id,
-        ]);
-
-        /** Can use `sortKeys()` on collection when Laravel >= 5.6 */
-        $all = $fields->all();
-        ksort($all);
-
-        return collect($all);
+        ])->sortKeys();
     }
 
     /**