+++ tests/lib/Integration/Eloquent/MorphToTest.php
@@ -74,8 +74,9 @@ class MorphToTest extends TestCase
 
         $id = $this
             ->actingAs($comment->user)
-            ->doCreate($data)
-            ->assertCreatedWithId($data);
+            ->doCreate($data, ['include' => 'created-by,commentable'])
+            ->assertCreatedWithServerId(url('/api/v1/comments'), $data)
+            ->id();
 
         $this->assertDatabaseHas('comments', [
             'id' => $id,
@@ -105,8 +106,9 @@ class MorphToTest extends TestCase
         ];
 
         $id = $this
-            ->doCreate($data)
-            ->assertCreatedWithId($data);
+            ->doCreate($data, ['include' => 'commentable'])
+            ->assertCreatedWithServerId(url('/api/v1/comments'), $data)
+            ->id();
 
         $this->assertDatabaseHas('comments', [
             'id' => $id,
@@ -130,7 +132,7 @@ class MorphToTest extends TestCase
             ],
         ];
 
-        $this->doUpdate($data)->assertUpdated($data);
+        $this->doUpdate($data, ['include' => 'commentable'])->assertUpdated($data);
 
         $this->assertDatabaseHas('comments', [
             'id' => $comment->getKey(),
@@ -160,7 +162,7 @@ class MorphToTest extends TestCase
             ],
         ];
 
-        $this->doUpdate($data)->assertUpdated($data);
+        $this->doUpdate($data, ['include' => 'commentable'])->assertUpdated($data);
 
         $this->assertDatabaseHas('comments', [
             'id' => $comment->getKey(),
@@ -190,7 +192,7 @@ class MorphToTest extends TestCase
             ],
         ];
 
-        $this->doUpdate($data)->assertUpdated($data);
+        $this->doUpdate($data, ['include' => 'commentable'])->assertUpdated($data);
 
         $this->assertDatabaseHas('comments', [
             'id' => $comment->getKey(),
@@ -215,7 +217,7 @@ class MorphToTest extends TestCase
         ];
 
         $this->doReadRelated($comment, 'commentable')
-            ->assertReadHasOne($expected);
+            ->assertFetchedOne($expected);
     }
 
     public function testReadRelatedNull()
@@ -224,7 +226,7 @@ class MorphToTest extends TestCase
         $comment = factory(Comment::class)->create();
 
         $this->doReadRelated($comment, 'commentable')
-            ->assertReadHasOne(null);
+            ->assertFetchedNull();
     }
 
     public function testReadRelationship()
@@ -232,7 +234,8 @@ class MorphToTest extends TestCase
         $comment = factory(Comment::class)->states('video')->create();
 
         $this->doReadRelationship($comment, 'commentable')
-            ->assertReadHasOneIdentifier('videos', $comment->commentable_id);
+            ->willSeeType('videos')
+            ->assertFetchedToOne($comment->commentable_id);
     }
 
     public function testReadEmptyRelationship()
@@ -240,7 +243,7 @@ class MorphToTest extends TestCase
         $comment = factory(Comment::class)->create();
 
         $this->doReadRelationship($comment, 'commentable')
-            ->assertReadHasOneIdentifier(null);
+            ->assertFetchedNull();
     }
 
     public function testReplaceNullRelationshipWithRelatedResource()