+++ tests/lib/Integration/Pagination/CursorPagingTest.php
@@ -91,7 +91,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(['page' => ['limit' => 4]])
-            ->assertSearchedIds($comments->take(4))
+            ->assertFetchedMany($comments->take(4))
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -127,7 +127,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -165,7 +165,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -206,7 +206,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -221,16 +221,14 @@ class CursorPagingTest extends TestCase
 
     /**
      * If the before key does not exist, we expect the cursor builder
-     * to throw an exception. Applications should validate the id
-     * before passing it to the cursor.
+     * to throw an exception which would constitute an internal server error.
+     * Applications should validate the id before passing it to the cursor.
      */
     public function testBeforeDoesNotExist()
     {
-        $this->expectException(\OutOfRangeException::class);
-        $this->expectExceptionMessage('999');
-
         $this->actingAsUser()
-            ->doSearch(['page' => ['before' => '999']]);
+            ->doSearch(['page' => ['before' => '999']])
+            ->assertStatus(500);
     }
 
     public function testAfter()
@@ -255,7 +253,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -292,7 +290,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -327,7 +325,7 @@ class CursorPagingTest extends TestCase
         $response = $this
             ->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -370,7 +368,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -413,7 +411,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -456,16 +454,14 @@ class CursorPagingTest extends TestCase
 
     /**
      * If the after key does not exist, we expect the cursor builder
-     * to throw an exception. Applications should validate the id
-     * before passing it to the cursor.
+     * to throw an exception which would constitute an internal server error.
+     * Applications should validate the id before passing it to the cursor.
      */
     public function testAfterDoesNotExist()
     {
-        $this->expectException(\OutOfRangeException::class);
-        $this->expectExceptionMessage('999');
-
         $this->actingAsUser()
-            ->doSearch(['page' => ['after' => '999']]);
+            ->doSearch(['page' => ['after' => '999']])
+            ->assertStatus(500);
     }
 
     /**
@@ -494,7 +490,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -537,7 +533,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [
@@ -578,7 +574,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'cursor' => [
@@ -618,7 +614,7 @@ class CursorPagingTest extends TestCase
 
         $this->actingAsUser()
             ->doSearch(compact('page'))
-            ->assertSearchedIds($expected)
+            ->assertFetchedMany($expected)
             ->assertJson([
                 'meta' => [
                     'page' => [